requirements.txt
Django==4.1.7
django-appconf==1.0.5
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7.1
sorl-thumbnail==12.9.0
sqlparse==0.4.3

python-dotenv~=0.21.1